Mlops/STUDY-BUDDY-AI/requirements.txt
langchain
langchain-groq
langchain-openai
langchain-core
pandas
streamlit
python-dotenv
pydantic
openai